requirements.txt
Django
email-validator
drf-yasg
django-rest-framework